+++ src/mcx_utils.h
@@ -184,8 +184,8 @@ typedef struct MCXConfig{
 #ifdef	__cplusplus
 extern "C" {
 #endif
-void mcx_savedata(float *dat, int len, Config *cfg);
-void mcx_savenii(float *dat, int len, char* name, int type32bit, int outputformatid, Config *cfg);
+void mcx_savedata(float *dat, size_t len, Config *cfg);
+void mcx_savenii(float *dat, size_t len, char* name, int type32bit, int outputformatid, Config *cfg);
 void mcx_error(const int id,const char *msg,const char *file,const int linenum);
 void mcx_loadconfig(FILE *in, Config *cfg);
 void mcx_saveconfig(FILE *in, Config *cfg);
